fix: deduped logical tree resolution
Ref SC-777

Found that the dependencies were being copied as a reference when the `copy` function was being called (and then overwritten), which results is a logical tree closer to what the `npm ls` output looks like.

This fix skips over the dependencies in the copy process and then rebuilds the deps. The important change is that the `from` property is correctly tracked when we're copying a full partial tree across from one package to another. The code also prevents circular (infinite) depth by tracking the heritage (the `from`) and if the package already appears in the path, it won't try to resolve the deps (again).

This can seen when you install the es6-map package as it has a large circular tree.