fix: pluck logic with varied paths from disk
Required a total refactor of code and logic in lib/pluck.js

The process now uses a similar method as the `require` process as if the root module were loaded, then incrementally loadedthe next module in the path. If it's found in the leaf's dependencies, then it moves on. If it's not found, it will go back up the tree of deps then the cycle back down (as seen in the refactored pluck function).